+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @ballerine/workflow-core
 
+## 0.6.89
+
+### Patch Changes
+
+- updated packages
+- Updated dependencies
+  - @ballerine/common@0.9.70
+
+## 0.6.88
+
+### Patch Changes
+
+- updated common and core
+- Updated dependencies
+  - @ballerine/common@0.9.69
+
+## 0.6.87
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.68
+
+## 0.6.86
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.67
+
 ## 0.6.85
 
 ### Patch Changes